requirements.txt
# web framework
django
djangorestframework

# data manipulation
pandas

# for postgres
psycopg2
